+++ src/features/picks/index.js
@@ -7,4 +7,9 @@ export { default as PicksSelfRecapSection } from './ui/PicksSelfRecapSection';
 export { default as usePicksForm } from './model/usePicksForm';
 export { usePicksSelfRecap } from './model/usePicksSelfRecap';
 export { useNextShowPicksStatus } from './model/useNextShowPicksStatus';
+export {
+  hasNonEmptyPicksObject,
+  pickEntryHasSubmission,
+  userHasSubmittedPickEntry,
+} from './model/pickSubmission';
 export { useSetlistLockToast } from './model/useSetlistLockToast';